+++ cookietemple/lint/domains/cli.py
@@ -15,7 +15,7 @@ class CliPythonLint(TemplateLinter, metaclass=GetLintingFunctionsMeta):
     def __init__(self, path):
         super().__init__(path)
 
-    def lint(self, is_create):
+    def lint(self, is_create, skip_external):
         super().lint_project(self, self.methods)
 
         # Call autopep8, if needed
@@ -24,6 +24,8 @@ class CliPythonLint(TemplateLinter, metaclass=GetLintingFunctionsMeta):
             autopep8 = Popen(['autopep8', self.path, '--recursive', '--in-place', '--pep8-passes', '2000'],
                              universal_newlines=True, shell=False, close_fds=True)
             (autopep8_stdout, autopep8_stderr) = autopep8.communicate()
+        elif skip_external:
+            pass
         elif cookietemple_questionary_or_dot_cookietemple(function='confirm',
                                                           question='Do you want to run autopep8 to fix pep8 issues?',
                                                           default='n'):